Transition to a single plugin class
