+++ packages/firebase_performance/android/src/main/java/io/flutter/plugins/firebaseperformance/FirebasePerformancePlugin.java
@@ -4,30 +4,72 @@
 
 package io.flutter.plugins.firebaseperformance;
 
+import android.util.SparseArray;
 import androidx.annotation.NonNull;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
 /** FirebasePerformancePlugin */
-public class FirebasePerformancePlugin implements FlutterPlugin {
+public class FirebasePerformancePlugin implements FlutterPlugin, MethodChannel.MethodCallHandler {
   private static final String CHANNEL_NAME = "plugins.flutter.io/firebase_performance";
 
   private MethodChannel channel;
 
   public static void registerWith(Registrar registrar) {
     final MethodChannel channel = new MethodChannel(registrar.messenger(), CHANNEL_NAME);
-    channel.setMethodCallHandler(new FirebasePerformanceHandler());
+    channel.setMethodCallHandler(new FirebasePerformancePlugin());
   }
 
   @Override
   public void onAttachedToEngine(FlutterPluginBinding binding) {
     channel = new MethodChannel(binding.getFlutterEngine().getDartExecutor(), CHANNEL_NAME);
-    channel.setMethodCallHandler(new FirebasePerformanceHandler());
+    channel.setMethodCallHandler(this);
   }
 
   @Override
   public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     channel.setMethodCallHandler(null);
+  }
+
+  private final SparseArray<MethodChannel.MethodCallHandler> handlers = new SparseArray<>();
+
+  void addHandler(final int handle, final MethodChannel.MethodCallHandler handler) {
+    if (handlers.get(handle) != null) {
+      final String message = String.format("Object for handle already exists: %s", handle);
+      throw new IllegalArgumentException(message);
+    }
+
+    handlers.put(handle, handler);
+  }
+
+  void removeHandler(final int handle) {
+    handlers.remove(handle);
+  }
+
+  private MethodChannel.MethodCallHandler getHandler(final MethodCall call) {
+    final Integer handle = call.argument("handle");
+
+    if (handle == null) return null;
+    return handlers.get(handle);
+  }
+
+  @Override
+  public void onMethodCall(MethodCall call, MethodChannel.Result result) {
+    if (call.method.equals("FirebasePerformance#instance")) {
+      handlers.clear();
+      final Integer handle = call.argument("handle");
+      addHandler(handle, new FlutterFirebasePerformance(this));
+      result.success(null);
+    } else {
+      final MethodChannel.MethodCallHandler handler = getHandler(call);
+
+      if (handler != null) {
+        handler.onMethodCall(call, result);
+      } else {
+        result.notImplemented();
+      }
+    }
   }
 }
